requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
djangorestframework-simplejwt==5.3.1
redis==5.0.1
drf-yasg==1.21.7
django-redis==5.4.0
python-dateutil==2.8.2
